requirements.txt
openai
numpy
pandas
python-dotenv
flask
